pyproject.toml
[tool.poetry]
name = "autoresearcher"
version = "0.1.0"
description = "AutoResearcher – medical RAG assistant"
authors = ["Sanjay <sanjaydv@umd.edu>"]
packages = [{ include = "autoresearcher", from = "src" }]

### ───────────────────────── Runtime deps ─────────────────────────
[tool.poetry.dependencies]
python = ">=3.10,<3.13"           # 3.12.8 included

# ── API / web ─────────────────────────────────────────────────────
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }

# ── Core typing / settings ────────────────────────────────────────
pydantic = "^2.7.3"
pydantic-settings = "^2.2.1"
numpy = "^1.26.4"

# ── Data ingestion & processing ───────────────────────────────────
lxml = "^5.2.1"                   # robust XML parsing (PMC)
pandas = "^2.2.2"
pyarrow = "^16.1.0"               # Parquet I/O

# ── Vector search / embeddings ────────────────────────────────────
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.7.0"  # local embeddings fallback

# ── LLM & fine‑tuning stack ───────────────────────────────────────
transformers = { extras = ["torch"], version = "^4.43.0" }
torch = "^2.3.0"                  # supports Python 3.12 wheels
accelerate = "^0.27.0"
peft = "^0.10.0"                  # LoRA / QLoRA adapters
trl = "^0.10.0"                   # SFTTrainer & PPO
datasets = "^2.19.0"

# ── Cloud / GCP integration ───────────────────────────────────────
google-cloud-aiplatform = "^1.44.0"   # Vertex AI embeddings & training
google-cloud-storage    = "^2.16.0"

# ── Networking / logging ──────────────────────────────────────────
httpx      = "^0.27.0"
structlog  = "^24.1.0"
python-dotenv = "^1.0.1"              # local .env loading

# ── Observability ─────────────────────────────────────────────────
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-api  = "^1.25.0"
opentelemetry-sdk  = "^1.25.0"

# ── Orchestration / jobs ──────────────────────────────────────────
prefect = "^2.18.2"

# ── Utilities ─────────────────────────────────────────────────────
tqdm  = "^4.66.4"
rich  = "^13.7.1"

### ───────────────────────── Dev / test deps ──────────────────────
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff   = "^0.4.4"
black  = "^24.4.2"
isort  = "^5.13.2"
mypy   = "^1.10.0"
types-requests = "^2.31.0.20240406"

### ───────────────────────── Build system ─────────────────────────
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
